add basic exception handling

+++ src/Database.py
@@ -1,14 +1,20 @@
 import json
+import os.path
 
 
 class Database:
 
     def __init__(self):
         self.db_location = "my_db.db"
-        with open(self.db_location, 'w+') as my_db:
-            json.dump({"000000_____TTTT__________": "X"}, my_db, indent=4)
-            my_db.seek(0, 0)
-            self.db: dict = json.load(my_db)
+        if not os.path.isfile(self.db_location):
+            with open(self.db_location, 'w+') as my_db:
+                # some key and value are needed for json to read the file
+                json.dump({"__DEFAULT_KEY__": "__DEFAULT_VALUE__"}, my_db, indent=4)
+                my_db.seek(0, 0)
+                self.db: dict = json.load(my_db)
+        else:
+            with open(self.db_location) as my_db:
+                self.db: dict = json.load(my_db)
 
     def __int__(self, location: str):
         self.db_location = location
@@ -18,6 +24,8 @@ class Database:
     def write_db_to_file(self) -> None:
         with open(self.db_location, 'r+') as my_file:
             json.dump(self.db, my_file, indent=4)
+
+            # If we don't seek to start, the json won't be able to read since after end, it's all empty
             my_file.seek(0, 0)
 
     def read_file(self, location) -> dict:
@@ -28,20 +36,24 @@ class Database:
     ### --------------------------------------------------------------------- ###
 
     def get(self, key: str):
-        print("get", key)
-        print(self.db[key])
+        # print("get", key)
+        try:
+            print(f'{key}={self.db[key]}')
+        except KeyError:
+            print(f"Key {key} does not exist in DB")
 
     def put(self, pair: str):
-        print("PUT ", pair)
+        # print("PUT ", pair)
         key, value = pair.split("=")
         self.db = self.db | json.loads(f'{{"{key}": "{value}"}}')
         self.write_db_to_file()
+        print(f"Successfully put {key}={value} in Database")
 
-    def put_from_file(self, file_path: str, db: dict):
-        self.db = db | self.read_file(file_path)
+    def put_from_file(self, file_path: str):
+        self.db = self.db | self.read_file(file_path)
         self.write_db_to_file()
 
     def delete(self, key: str):
-        print("delete ", key)
+        # print("delete ", key)
         self.db.pop(key)
         self.write_db_to_file()
